+++ demo/app/app.module.ts
@@ -27,6 +27,7 @@ import { MaxValidatorDirective } from './shared/settings-form/max-validtor.direc
 import { MinValidatorDirective } from './shared/settings-form/min-validator.directive';
 import { MapsLayerComponent } from './components/maps-layer/maps-layer.component';
 import { ModelsLayerComponent } from './components/models-layer/models-layer.component';
+import { PolygonPerformanceTestComponent } from './components/polygon-layer/polygon-performance-test/polygon-performance-test.component';
 import { KeyboardControlLayerComponent } from './components/keyboard-control-layer/keyboard-control-layer.component';
 
 
@@ -52,6 +53,7 @@ import { KeyboardControlLayerComponent } from './components/keyboard-control-lay
 		ModelsLayerComponent,
     MaxValidatorDirective,
     MinValidatorDirective,
+		PolygonPerformanceTestComponent,
     KeyboardControlLayerComponent,
   ],
   imports: [