+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.0.15
 ### Features
 * Expose `MapEventManagerService` from `<ac-map #map/>` component reference: `acMapComponent.getMapEventManager()`
-
+* Add drag and drop event to `MapEventManagerService`
 
 ## 0.0.14
 ### Breaking changes